% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_step.R
\name{e_step_general_impute}
\alias{e_step_general_impute}
\title{General E-step Imputation}
\usage{
e_step_general_impute(data, mu, Sigma)
}
\arguments{
\item{data}{A numeric matrix with NAs}

\item{mu}{Mean vector}

\item{Sigma}{Covariance matrix}
}
\value{
Matrix with missing entries imputed using conditional expectations
}
\description{
General E-step Imputation
}
