% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_step_poisson_em.R
\name{e_step_poisson_em}
\alias{e_step_poisson_em}
<<<<<<< HEAD
\title{Standard E-step Imputation}
=======
\title{Standard E-step Imputation (Poisson)}
>>>>>>> f1bc9dbc6ff6d4895791766c65ddbe29c6d8a016
\usage{
e_step_poisson_em(data, params)
}
\arguments{
\item{data}{A numeric matrix with missing values (NAs).}

<<<<<<< HEAD
\item{lambda}{A numeric vector representing the mean of the distribution.}
=======
\item{lambda}{A numeric vector representing the lambda of the distribution.}
>>>>>>> f1bc9dbc6ff6d4895791766c65ddbe29c6d8a016
}
\value{
A numeric matrix of the same dimensions as \code{data}, with missing
values imputed using the conditional expectation given observed values.
}
\description{
Performs the E-step of the EM algorithm by imputing missing entries in a
<<<<<<< HEAD
numeric matrix using conditional expectations under the poisson distribution model.
=======
numeric matrix using conditional expectations under the poisson model.
>>>>>>> f1bc9dbc6ff6d4895791766c65ddbe29c6d8a016
}
\details{
This function is typically called internally by the EM loop
(\code{\link{run_em_algorithm}}), but it can be used standalone for testing.
}
\examples{
data <- matrix(c(1, NA, 3, 4), ncol = 2)
lambda <- colMeans(data, na.rm = TRUE)
params <- list(lambda = lambda)
e_step_poisson_em(data, params)
}
\seealso{
<<<<<<< HEAD
\code{\link{run_em_algorithm}}, \code{\link{log_likelihood_poisson}}
=======
\code{\link{run_em_algorithm}}
>>>>>>> f1bc9dbc6ff6d4895791766c65ddbe29c6d8a016
}
