% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{grapes_or_or_grapes}
\alias{grapes_or_or_grapes}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand value}

\item{y}{Default value}
}
\value{
The left-hand value if not NULL; otherwise the right-hand value.
}
\description{
Returns the left-hand side if it is not NULL, otherwise returns the right-hand side.
}
\examples{
NULL \%||\% "default"
1 \%||\% "ignored"

}
