% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_parameters.R
\name{initialize_parameters}
\alias{initialize_parameters}
\title{Dispatcher to initialize parameters based on distribution}
\usage{
initialize_parameters(data, dist)
}
\arguments{
\item{data}{Matrix with missing values}

\item{dist}{Character string: "mvnorm", "poisson", "mixture", etc.}
}
\value{
A list of initial parameters (depends on distribution)
}
\description{
Dispatcher to initialize parameters based on distribution
}
