% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_step.R
\name{m_step_estimate}
\alias{m_step_estimate}
\title{Compute third moment from the mean (variation B)}
\usage{
m_step_estimate(imputed_data)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A single numeric value
}
\description{
Another version for computing a third-order central moment.
}
