% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_step_poisson.R
\name{m_step_poisson}
\alias{m_step_poisson}
\title{M-step for Poisson Model}
\usage{
m_step_poisson(imputed_data)
}
\arguments{
\item{imputed_data}{A numeric matrix where missing values have been filled in,
typically using conditional expectations from the E-step.}
}
\value{
A named list containing:
\describe{
\item{lambda}{A numeric vector of araveage rate of events}
}
}
\description{
Performs the M-step of the EM algorithm for a poisson distribution,
given a matrix with missing values imputed (i.e., the completed dataset).
}
\details{
This function calls a C++ routine (\code{_DATA501Package_m_step_poisson})
to compute the MLEs for the poisson parameters \eqn{(\lambda)}.
It is designed to work with both outputs of \code{\link{e_step_poisson}}.
}
\examples{
data <- matrix(c(1, NA, 3, 4), ncol = 2)
lambda <- colMeans(data, na.rm = TRUE)
params <- list(lambda = lambda)
imputed <- e_step_poisson(data, params)
m_step_poisson(imputed)
}
\seealso{
\code{\link{e_step_poisson}}, \code{\link{em_engine}}, \code{\link{log_likelihood_poisson}}
}
