% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_model_class.R
\name{em_model}
\alias{em_model}
\title{Constructor for em_model object}
\usage{
em_model(data, method = "EM", early_stop = list())
}
\arguments{
\item{data}{Input data matrix with missing values}

\item{method}{"EM" or "MCEM"}

\item{early_stop}{List of early stopping criteria}
}
\value{
An object of class em_model
}
\description{
Constructor for em_model object
}
