% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_parameters.R
\name{initialize_parameters}
\alias{initialize_parameters}
\title{Initialize parameters from incomplete data}
\usage{
initialize_parameters(data)
}
\arguments{
\item{data}{A numeric matrix with NAs}
}
\value{
A list with estimated \code{mu} and \code{Sigma}
}
\description{
Initialize parameters from incomplete data
}
