% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_em_algorithm.R
\name{run_em_algorithm}
\alias{run_em_algorithm}
\title{Run EM or MCEM depending on model$method}
\usage{
run_em_algorithm(
  model,
  tolerance = 1e-05,
  max_iter = 100,
  m = NULL,
  burn = NULL,
  thin = NULL,
  tau = NULL
)
}
\arguments{
\item{model}{An object of class \code{em_model}, containing the data, method ("EM" or "MCEM"),
and distribution type.}

\item{tolerance}{Numeric. Threshold for convergence (change in log-likelihood). Default is \code{1e-5}.}

\item{max_iter}{Integer. Maximum number of EM iterations. Default is \code{100}.}

\item{m}{Integer. Number of MCMC samples (used only if \code{method == "MCEM"}).}

\item{burn}{Integer. Burn-in samples for MCMC (MCEM only).}

\item{thin}{Integer. Thinning factor for MCMC (MCEM only).}

\item{tau}{Numeric. Proposal scale for Metropolis-Hastings (MCEM only).}
}
\value{
An updated \code{em_model} object with filled-in fields such as \code{loglik_history},
\code{parameters}, \code{imputed}, and \code{early_stop}.
}
\description{
In this draft package: only multivariate normal distribution is ready for testing (method = 'nvnorm')
Executes the Expectation-Maximization (EM) or Monte Carlo EM (MCEM) algorithm
depending on the \code{method} field of the input \code{em_model} object.
}
\examples{
data <- matrix(c(1, NA, 3, 4), ncol = 2)
model <- em_model(data, method = "EM")
result <- run_em_algorithm(model, max_iter = 10)
result$imputed
}
\seealso{
\code{\link{em_engine}}, \code{\link{em_model}}, \code{\link{e_step_nvnorm}}
}
